backend/requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
sqlalchemy==2.0.48
pydantic==2.12.5
python-dotenv==1.2.2
python-multipart==0.0.22
PyJWT==2.12.1
bcrypt==5.0.0
passlib==1.7.4
google-auth==2.38.0
requests==2.32.3
slowapi==0.1.9
gunicorn==23.0.0
psycopg2-binary==2.9.10
redis[hiredis]==5.2.1
websockets==15.0.1
alembic==1.14.1
celery==5.4.0
cryptography==44.0.2
tenacity==9.1.2
